api_ads/pyproject.toml
[tool.poetry]
name = "api-ads"
version = "0.1.0"
description = ""
authors = ["Andrei Emelianenko <spliitvein@gmail.com>"]
readme = "README.md"
packages = [{include = "api_ads"}]


[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
whitenoise = "^6.6.0"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
dj-database-url = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
